refactor(userspace-dp): split afxdp/umem/tests.rs into per-concern files (#4667)
Pure test-code motion. The umem tests lived in a single 1765-line
tests.rs loaded via `#[path = "tests.rs"] mod tests;`, over the
modularity-discipline LOC threshold. Split it into a tests/ subdir with
per-concern sibling submodules, matching the #4664/#4665/#4651
convention (a tests/mod.rs holding the shared `use` header plus the one
cross-concern fixture, each concern reached via `use super::*`).

Concern files and their test counts:
  mmap_area.rs             (1) MmapArea slice-bounds beyond registered len
  tx_inbox.rs              (3) redirect-inbox admission + overflow counters
  latency_buckets.rs       (6) bucket_index_for_ns layout + drain /
                               redirect-acquire histogram sampling + seeding
  snapshot_propagation.rs  (3) snapshot() owner-profile / drop-counter
                               propagation + owner-profile cacheline isolation
  tx_submit_latency.rs     (9) #812 TX submit-latency stamp/reap fold, OOB
                               handling, single- + cross-thread skew bound
  tx_kick_latency.rs       (6) #825 TX kick-latency bucket mapping,
                               accumulation, sentinels, cross-thread skew
  debug_state.rs           (5) #943 v_min scratch flush + idle debug-state
                               cadence / active-flow aging (+ local fixtures)

The shared `use` header (ACTIVE_WINDOW_EPOCHS, EqualFlowTargetPolicy,
`use super::*`) and the one cross-concern fixture
(`test_tx_request_for_inbox`, used by tx_inbox and latency_buckets) live
in tests/mod.rs; concern-local helpers travel with their tests. The
new tests/mmap_area.rs / tests/debug_state.rs names are distinct from
the pre-existing production umem/mmap.rs / umem/debug_state.rs (separate
module path — no collision), and from the co-located mmap.rs
harden-tests in umem/mmap_tests.rs.

The parent umem/mod.rs declaration changed from
`#[path = "tests.rs"] mod tests;` to plain `mod tests;` (resolving to
tests/mod.rs). The only other production-file edits are comment-only:
two #1351 rationale blocks referenced `umem/tests.rs:NNNN` line numbers
that would dangle after the split, repointed to
`umem/tests/debug_state.rs` (no logic change). README.md's file table
gains the tests/ subdir entry.

No production code and no test logic changed. Every test/helper body
moved byte-identical (extracted via an exact line-range partition of the
original file; all items are top-level so no dedent was needed). A
code-line multiset check confirms zero original code lines are missing
and the only new lines are the seven `mod X;` decls plus the seven
per-file `use super::*;` wiring lines.

Validation:
  - test count identical: 33 umem tests before and after; whole
    userspace-dp/src #[test]/#[tokio::test] count 3336 == 3336.
  - code-line multiset: missing=0; extra = 7x `mod X;` + 7x
    `use super::*;` (wiring only).
  - cargo build clean (pre-existing crate warnings only).
  - full cargo test suite (release, --test-threads=1): 3860 passed,
    0 failed, 2 ignored (EXIT=0) across the crate's test binaries
    (60 + 3769 + 8 + 22 + 1). The 33 moved umem tests all execute
    (`afxdp::umem::tests::` --list == 33) and pass.

Closes #4667
